+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Warning message instead of histogram when all values are 0, to avoid confusing display behaviour: https://github.com/riparias/gbif-alert/issues/92
 - Fix a minor, recently introduced display issue: https://github.com/riparias/gbif-alert/issues/255
+- Unused datasets are automatically cleaned up at import time: https://github.com/riparias/gbif-alert/issues/222
 
 # v1.3.0 (2023-08-30)